+++ Assignment1/POCO/Account.cs
@@ -5,9 +5,9 @@
         public int AccountNumber { get; set; }
         public char AccountType { get; set; }
         public int CustomerId { get; set; }
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
 
-        public Account(int accountNumber, char accountType, int customerId, double balance)
+        public Account(int accountNumber, char accountType, int customerId, decimal balance)
         {
             AccountNumber = accountNumber;
             AccountType = accountType;
